useThemeSeriesColors instead of directly applying colors - extract useThemeSeriesColors to global hooks - use DataSeries instead of local Series type
