+++ src/components/MultiSeriesBarChart/Chart.tsx
@@ -24,7 +24,6 @@ import type {
   Series,
   RenderTooltipContentData,
   BarOptions as MultiSeriesBarOptions,
-  GridOptions,
   XAxisOptions,
   YAxisOptions,
 } from './types';
@@ -43,8 +42,8 @@ interface Props {
   chartDimensions: Dimensions;
   renderTooltipContent(data: RenderTooltipContentData): React.ReactNode;
   barOptions: BarOptions;
-  gridOptions: GridOptions & GridTheme;
-  xAxisOptions: XAxisOptions & XAxisTheme;
+  gridOptions: Omit<GridTheme, 'showVerticalLines'>;
+  xAxisOptions: Omit<XAxisOptions & XAxisTheme, 'hide'>;
   yAxisOptions: YAxisOptions & YAxisTheme;
   isAnimated?: boolean;
   emptyStateText?: string;
