+++ src/components/MultiSeriesBarChart/types.ts
@@ -49,13 +49,6 @@ export interface BarOptions {
   zeroAsMinHeight: boolean;
 }
 
-export interface GridOptions {
-  showHorizontalLines: boolean;
-  color: string;
-  horizontalOverflow: boolean;
-  horizontalMargin: number;
-}
-
 export interface XAxisOptions {
   labelFormatter: StringLabelFormatter;
   labels: string[];
