+++ src/components/StackedAreaChart/hooks/index.ts
@@ -1,3 +1,2 @@
 export {useXScale} from './use-x-scale';
 export {useYScale} from './use-y-scale';
-export {useThemeSeriesColors} from './use-theme-series-colors';
