+++ src/hooks/index.ts
@@ -6,3 +6,4 @@ export {usePrevious} from './use-previous';
 export {useResizeObserver} from './useResizeObserver';
 export {useTheme} from './useTheme';
 export {usePolarisVizContext} from './usePolarisVizContext';
+export {useThemeSeriesColors} from './use-theme-series-colors';
