+++ src/hooks/tests/use-theme-series-colors.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
 
-import type {Theme} from '../../../../types';
+import type {Theme} from '../../types';
 import {
   useThemeSeriesColors,
   getSeriesColors,
